+++ Source/Tarkov-QuestTree/ModInfo.cs
@@ -13,7 +13,7 @@ namespace QuestTree
     /// </summary>
     internal static class ModInfo
     {
-        public const string Version = "1.13.3";
+        public const string Version = "1.14.0";
 
         private static string _stamp;